backend/requirements.txt
anyio==3.6.2
autopep8==2.0.2
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3
click-default-group-wheel==1.2.2
cryptography==40.0.2
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
fastapi==0.95.1
h11==0.14.0
httptools==0.5.0
idna==3.4
passlib==1.7.4
pyasn1==0.5.0
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.7
PyJWT==2.6.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-jose==3.3.0
PyYAML==6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.10
sqlite-fts4==1.0.3
sqlite-utils==3.30
starlette==0.26.1
tabulate==0.9.0
typing_extensions==4.5.0
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.2
